+++ agent/plugin/plugin_test.go
@@ -400,37 +400,18 @@ func TestConfigurationToEnvironment(t *testing.T) {
 			configJSON: `{ "and _ with a    - number": 12 }`,
 			wantEnvMap: map[string]string{
 				"BUILDKITE_PLUGIN_CONFIGURATION":                           `{"and _ with a    - number":12}`,
-				"BUILDKITE_PLUGIN_DOCKER_COMPOSE_AND_WITH_A_NUMBER":        "12",
 				"BUILDKITE_PLUGIN_DOCKER_COMPOSE_AND___WITH_A______NUMBER": "12",
 				"BUILDKITE_PLUGIN_NAME":                                    "DOCKER_COMPOSE",
 			},
-			expectedErr: (&DeprecatedNameErrors{}).Append(
-				DeprecatedNameError{
-					old: "BUILDKITE_PLUGIN_DOCKER_COMPOSE_AND_WITH_A_NUMBER",
-					new: "BUILDKITE_PLUGIN_DOCKER_COMPOSE_AND___WITH_A______NUMBER",
-				},
-			),
 		},
 		{
 			configJSON: `{ "and _ with a    - number": 12, "A - B": 13 }`,
 			wantEnvMap: map[string]string{
 				"BUILDKITE_PLUGIN_CONFIGURATION":                           `{"A - B":13,"and _ with a    - number":12}`,
-				"BUILDKITE_PLUGIN_DOCKER_COMPOSE_AND_WITH_A_NUMBER":        "12",
 				"BUILDKITE_PLUGIN_DOCKER_COMPOSE_AND___WITH_A______NUMBER": "12",
-				"BUILDKITE_PLUGIN_DOCKER_COMPOSE_A_B":                      "13",
 				"BUILDKITE_PLUGIN_DOCKER_COMPOSE_A___B":                    "13",
 				"BUILDKITE_PLUGIN_NAME":                                    "DOCKER_COMPOSE",
 			},
-			expectedErr: (&DeprecatedNameErrors{}).Append(
-				DeprecatedNameError{
-					old: "BUILDKITE_PLUGIN_DOCKER_COMPOSE_AND_WITH_A_NUMBER",
-					new: "BUILDKITE_PLUGIN_DOCKER_COMPOSE_AND___WITH_A______NUMBER",
-				},
-				DeprecatedNameError{
-					old: "BUILDKITE_PLUGIN_DOCKER_COMPOSE_A_B",
-					new: "BUILDKITE_PLUGIN_DOCKER_COMPOSE_A___B",
-				},
-			),
 		},
 		{
 			configJSON: `{ "bool-true-key": true, "bool-false-key": false }`,