+++ clicommand/meta_data_get.go
@@ -6,7 +6,7 @@ import (
 	"slices"
 	"time"
 
-	"github.com/buildkite/agent/v3/api"
+	"github.com/buildkite/agent/v4/api"
 	"github.com/buildkite/roko"
 	"github.com/urfave/cli"
 )
@@ -102,15 +102,14 @@ var MetaDataGetCommand = cli.Command{
 					cfg.Key,
 					cfg.Default,
 				)
-				fmt.Fprint(c.App.Writer, cfg.Default)
+				fmt.Fprintln(c.App.Writer, cfg.Default)
 				return nil
 			}
 
 			return fmt.Errorf("failed to get meta-data: %w", err)
 		}
 
-		// TODO: in the next agent magor version, we should terminate with a newline using fmt.FPrintln
-		_, err = fmt.Fprint(c.App.Writer, metaData.Value)
+		_, err = fmt.Fprintln(c.App.Writer, metaData.Value)
 		return err
 	},
 }