+++ internal/job/checkout.go
@@ -1,6 +1,7 @@
 package job
 
 import (
+	"cmp"
 	"context"
 	"errors"
 	"fmt"
@@ -11,11 +12,11 @@ import (
 	"strings"
 	"time"
 
-	"github.com/buildkite/agent/v3/internal/experiments"
-	"github.com/buildkite/agent/v3/internal/osutil"
-	"github.com/buildkite/agent/v3/internal/self"
-	"github.com/buildkite/agent/v3/internal/shell"
-	"github.com/buildkite/agent/v3/tracetools"
+	"github.com/buildkite/agent/v4/internal/experiments"
+	"github.com/buildkite/agent/v4/internal/osutil"
+	"github.com/buildkite/agent/v4/internal/self"
+	"github.com/buildkite/agent/v4/internal/shell"
+	"github.com/buildkite/agent/v4/tracetools"
 	"github.com/buildkite/roko"
 	"github.com/buildkite/shellwords"
 )
@@ -189,16 +190,24 @@ func (e *Executor) CheckoutPhase(ctx context.Context) error {
 	}
 
 	// Run post-checkout hooks
-	if err := e.executeGlobalHook(ctx, "post-checkout"); err != nil {
-		return err
-	}
-
-	if err := e.executeLocalHook(ctx, "post-checkout"); err != nil {
-		return err
-	}
-
-	if err := e.executePluginHook(ctx, "post-checkout", e.pluginCheckouts); err != nil {
-		return err
+	if experiments.IsEnabled(ctx, experiments.LegacyPostHookOrder) {
+		// Legacy order = same order as pre-checkout
+		if err := cmp.Or(
+			e.executeGlobalHook(ctx, "post-checkout"),
+			e.executeLocalHook(ctx, "post-checkout"),
+			e.executePluginHook(ctx, "post-checkout", e.pluginCheckouts),
+		); err != nil {
+			return err
+		}
+	} else {
+		// Modern order = reverse, to make composition easier
+		if err := cmp.Or(
+			e.executePluginHook(ctx, "post-checkout", e.pluginCheckoutsReversed),
+			e.executeLocalHook(ctx, "post-checkout"),
+			e.executeGlobalHook(ctx, "post-checkout"),
+		); err != nil {
+			return err
+		}
 	}
 
 	// Capture the new checkout path so we can see if it's changed.
@@ -1036,10 +1045,7 @@ func (e *Executor) defaultCheckoutPhase(ctx context.Context) error {
 	}
 
 	// resolve BUILDKITE_COMMIT based on the local git repo
-	if experiments.IsEnabled(ctx, experiments.ResolveCommitAfterCheckout) {
-		e.shell.Commentf("Using resolve-commit-after-checkout experiment 🧪")
-		e.resolveCommit(ctx)
-	}
+	e.resolveCommit(ctx)
 
 	return nil
 }